+++ Photo Club Hub/ViewModel/ListReaders/Level1JsonReader.swift
@@ -21,55 +21,7 @@ private let fileSubType = "level1" // level1 is part of file name, not the exten
 private let fileType = "json"
 private let organizationTypesToLoad: [OrganizationTypeEnum] = [.club, .museum]
 
-/* Example of a minimal level1.json file
-{
-    "clubs": [
-        {
-            "idPlus": {
-                "town": "Eindhoven",
-                "fullName": "Fotogroep de Gender",
-                "nickName": "FG deGender"
-            },
-            "coordinates": {
-                "latitude": 51.42398,
-                "longitude": 5.45010
-            }
-            "optional": {
-                "website": "https://www.fcdegender.nl",
-                "level2URL": "https://www.example.com/fgDeGender.level2.json"
-                "remark": [
-                    {
-                        "language": "NL",
-                        "value": "In dit museum zijn scenes van het TV programma 'Het Perfecte Plaatje' opgenomen."
-                    }
-                ]
-            }
-        },
-    "museums": [
-        {
-            "idPlus": {
-                "town": "New York",
-                "fullName": "Fotografiska New York",
-                "nickName": "Fotografiska NYC"
-            },
-            "coordinates": {
-                "latitude": 40.739278,
-                "longitude": -73.986722
-            }
-            "optional:" {
-                "website": "https://www.fotografiska.com/nyc/",
-                "wikipedia": "https://en.wikipedia.org/wiki/Fotografiska_New_York",
-                "remark": [
-                    {
-                        "language": "EN",
-                        "value": "Associated with the original Fotografiska Museum in Stockholm"
-                    }
-                ]
-            }
-        }
-    ]
-}
-*/
+// see xampleMin.level1.json and xampleMax.level1.json for syntax examples
 
 class Level1JsonReader {